refactor: Use OrderType instead of TypeOption
